third_party/WebKit/LayoutTests/fast/dom/custom/document-register-constructor.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function () {
    var proto = Object.create(HTMLElement.prototype);
    var ctor = document.registerElement('x-a', {prototype: proto});
    assert_true(ctor instanceof Function, 'constructor must be a function');
    assert_equals(typeof ctor, 'function', 'constructor must be a function instance');
}, 'constructor type');

test(function () {
    var proto = Object.create(HTMLElement.prototype);
    var ctor = document.registerElement('x-b', {prototype: proto});

    // FIXME: These are not specified yet. Update these assertions
    // when the name is specified.
    assert_equals(ctor.name, 'x-b', 'the constructor\'s name should match the type');
}, 'constructor name');

test(function () {
    var proto = Object.create(HTMLElement.prototype);
    var ctor = document.registerElement('x-c', {prototype: proto});
    assert_own_property(proto, 'constructor', 'document.register must configure the constructor property of the prototype');

    assert_equals(proto.constructor, ctor, 'the value of the constructor property must be the constructor function');

    var desc = Object.getOwnPropertyDescriptor(proto, 'constructor');
    assert_true(desc.writable, 'constructor property must be writable');
    assert_false(desc.enumerable, 'constructor property must be non-enumerable');
    assert_true(desc.configurable, 'constructor property must be configurable');
}, 'prototype "constructor" property');

test(function () {
    var proto = Object.create(HTMLElement.prototype);
    var ctor = document.registerElement('x-d', {prototype: proto});
    assert_own_property(ctor, 'prototype', 'document.register must configure the prototype property of the constructor');

    assert_equals(ctor.prototype, proto, 'the value of the prototype property must be the prototype object');

    var desc = Object.getOwnPropertyDescriptor(ctor, 'prototype');
    assert_false(desc.writable, 'prototype property must not be writable');
    assert_false(desc.enumerable, 'prototype property must be non-enumerable');
    assert_false(desc.configurable, 'prototype property must not be configurable');
}, 'constructor "prototype" property');
</script>
